# openapi.json
{"openapi":"3.1.0","info":{"title":"West Texas Sports Grill API","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/menu/categories":{"get":{"summary":"Get Categories","operationId":"get_categories_api_menu_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/menu/items":{"get":{"summary":"Get Menu Items","operationId":"get_menu_items_api_menu_items_get","parameters":[{"name":"featured","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Featured"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Menu Item","operationId":"create_menu_item_api_menu_items_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuItemCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/menu/items/{item_id}":{"put":{"summary":"Update Menu Item","operationId":"update_menu_item_api_menu_items__item_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuItemCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Menu Item","operationId":"delete_menu_item_api_menu_items__item_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/register":{"post":{"summary":"Register","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/":{"get":{"summary":"List Orders","operationId":"list_orders_api_orders__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Create Order","operationId":"create_order_api_orders__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/{order_id}/status":{"put":{"summary":"Update Order Status","operationId":"update_order_status_api_orders__order_id__status_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/track/{order_id}":{"get":{"summary":"Track Order","operationId":"track_order_api_orders_track__order_id__get","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/":{"get":{"summary":"List Reviews","operationId":"list_reviews_api_reviews__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Create Review","operationId":"create_review_api_reviews__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/pending":{"get":{"summary":"Pending Reviews","operationId":"pending_reviews_api_reviews_pending_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/reviews/{review_id}/approve":{"put":{"summary":"Approve Review","operationId":"approve_review_api_reviews__review_id__approve_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"integer","title":"Review Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{review_id}":{"delete":{"summary":"Delete Review","operationId":"delete_review_api_reviews__review_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"integer","title":"Review Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reservations/available-times":{"get":{"summary":"Available Times","operationId":"available_times_api_reservations_available_times_get","parameters":[{"name":"date","in":"query","required":true,"schema":{"type":"string","title":"Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reservations/":{"get":{"summary":"List Reservations","operationId":"list_reservations_api_reservations__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Create Reservation","operationId":"create_reservation_api_reservations__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reservations/{reservation_id}/status":{"put":{"summary":"Update Reservation Status","operationId":"update_reservation_status_api_reservations__reservation_id__status_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"reservation_id","in":"path","required":true,"schema":{"type":"integer","title":"Reservation Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/dashboard":{"get":{"summary":"Admin Dashboard","operationId":"admin_dashboard_api_admin_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/settings/":{"get":{"summary":"Get Settings","operationId":"get_settings_api_settings__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/settings/upload-image/{slot}":{"post":{"summary":"Upload Image","operationId":"upload_image_api_settings_upload_image__slot__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"slot","in":"path","required":true,"schema":{"type":"string","title":"Slot"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_image_api_settings_upload_image__slot__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/settings/image/{slot}":{"delete":{"summary":"Delete Image","operationId":"delete_image_api_settings_image__slot__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"slot","in":"path","required":true,"schema":{"type":"string","title":"Slot"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"user":{"additionalProperties":true,"type":"object","title":"User"}},"type":"object","required":["access_token","user"],"title":"AuthResponse"},"Body_login_api_auth_login_post":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"Body_login_api_auth_login_post"},"Body_upload_image_api_settings_upload_image__slot__post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_image_api_settings_upload_image__slot__post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MenuItemCreate":{"properties":{"name":{"type":"string","title":"Name"},"category_id":{"type":"integer","title":"Category Id"},"price":{"type":"number","title":"Price"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"calories":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Calories"},"allergens":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Allergens"},"is_featured":{"type":"boolean","title":"Is Featured","default":false},"is_popular":{"type":"boolean","title":"Is Popular","default":false},"is_spicy":{"type":"boolean","title":"Is Spicy","default":false},"is_available":{"type":"boolean","title":"Is Available","default":true}},"type":"object","required":["name","category_id","price"],"title":"MenuItemCreate"},"OrderCreate":{"properties":{"order_type":{"type":"string","title":"Order Type"},"items":{"items":{"$ref":"#/components/schemas/OrderItemPayload"},"type":"array","title":"Items"},"special_instructions":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Special Instructions"},"delivery_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Delivery Address"},"guest_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Guest Name"},"guest_email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Guest Email"},"guest_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Guest Phone"}},"type":"object","required":["order_type","items"],"title":"OrderCreate"},"OrderItemPayload":{"properties":{"menu_item_id":{"type":"integer","title":"Menu Item Id"},"quantity":{"type":"integer","title":"Quantity"}},"type":"object","required":["menu_item_id","quantity"],"title":"OrderItemPayload"},"ReservationPayload":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"phone":{"type":"string","title":"Phone"},"party_size":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Party Size"},"date":{"type":"string","title":"Date"},"time":{"type":"string","title":"Time"},"special_requests":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Special Requests"}},"type":"object","required":["name","email","phone","party_size","date","time"],"title":"ReservationPayload"},"ReviewPayload":{"properties":{"name":{"type":"string","title":"Name"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"body":{"type":"string","title":"Body"},"rating":{"type":"integer","maximum":5.0,"minimum":1.0,"title":"Rating"}},"type":"object","required":["name","body","rating"],"title":"ReviewPayload"},"UserRegister":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","email","password"],"title":"UserRegister"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}